// src/types/user.ts
import { ConnectionStatusType } from './serverTypes';

export type UserInvoice = {
  id: string;
  createdAt: Date | string | number;
  price: number;
};

export type CreditCard = {
  id: string;
  cardNumber: string;
  cardType: string;
};

export type Follower = {
  id: string;
  avatarUrl: string;
  name: string;
  country: string;
  isFollowed: boolean;
};

export type Gallery = {
  id: string;
  title: string;
  postAt: Date | string | number;
  imageUrl: string;
};

export type UserAddressBook = {
  id: string;
  name: string;
  phone: string;
  country: string;
  state: string;
  city: string;
  street: string;
  zipCode: string;
};

export type Profile = {
  id: string;
  cover: string;
  position: string;
  follower: number;
  following: number;
  quote: string;
  country: string;
  email: string;
  company: string;
  school: string;
  role: string;
  facebookLink: string;
  instagramLink: string;
  linkedinLink: string;
  twitterLink: string;
};

export type UserManager = {
  id: string;
  avatarUrl: string;
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
  country: string;
  state: string;
  city: string;
  zipCode: string;
  company: string;
  isVerified: boolean;
  status: string;
  role: string;
};

export type UserData = {
  id: string;
  avatarUrl: string;
  cover: string;
  name: string;
  follower: number;
  following: number;
  totalPost: number;
  position: string;
};

export type NotificationSettings = {
  activityComments: boolean;
  activityAnswers: boolean;
  activityFollows: boolean;
  applicationNews: boolean;
  applicationProduct: boolean;
  applicationBlog: boolean;
};

export type Friend = {
  id: string;
  avatarUrl: string;
  name: string;
  role: string;
};

export type UserPost = {
  id: string;
  author: {
    id: string;
    avatarUrl: string;
    name: string;
  };
  isLiked: boolean;
  createdAt: Date | string | number;
  media: string;
  message: string;
  personLikes: {
    name: string;
    avatarUrl: string;
  }[];
  comments: {
    id: string;
    author: {
      id: string;
      avatarUrl: string;
      name: string;
    };
    createdAt: Date | string | number;
    message: string;
  }[];
};

export interface ISearchedUser {
  id: string;
  fullName?: string | null;
  avatarUrl?: string | null;
  userName?: string | null;
  headLine?: string | null;
}

export interface ISearchUserResponse {
  id?: any | null;
  coverUrl?: string | null;
  avatarUrl?: string | null;
  headline?: string | null;
  cognitoUserId?: string | null;
  userName?: string | null;
  fullName?: string | null;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  otherToMeStatus?: ConnectionStatusType | null;
  meToOtherStatus?: ConnectionStatusType | null;
}

export interface ISearchNgoReponse {
  id?: any | null;
  fullName?: string | null;
  avatarUrl?: string | null;
  otherToMeStatus?: ConnectionStatusType | null;
  meToOtherStatus?: ConnectionStatusType | null;
}